requirements.txt
-r src/requirements.txt

aws-cdk.core==1.127.0
aws_cdk.aws_lambda==1.127.0
aws_cdk.aws_lambda_python==1.127.0

pytest
pytest-cov

flake8<4
flake8-commas
flake8-quotes
